refactor(player): extract up-next trigger policy from the overlay widget
